Refactor separating logic from patch
