+++ CookedInfo/CookedInfo.csproj
@@ -9,9 +9,9 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>CookedInfo</RootNamespace>
     <AssemblyName>CookedInfo</AssemblyName>
-	<OutputPath>bin\Debug</OutputPath>
-	<ModName>CookedInfo</ModName>
-	<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <OutputPath>bin\Debug</OutputPath>
+    <ModName>CookedInfo</ModName>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -55,6 +55,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DescriptionBuilder.cs" />
     <Compile Include="Plugin.cs" />
     <Compile Include="Patches.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -62,7 +63,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)$(AssemblyName).VS.props" />
   <Target Name="AfterBuild">
-	<Delete Files="$(BepInExPluginsDirectory)\$(ModName).dll" />
-	<Copy SourceFiles="$(ProjectDir)\bin\Debug\$(ModName).dll" DestinationFolder="$(BepInExPluginsDirectory)" ContinueOnError="true" />
+    <Delete Files="$(BepInExPluginsDirectory)\$(ModName).dll" />
+    <Copy SourceFiles="$(ProjectDir)\bin\Debug\$(ModName).dll" DestinationFolder="$(BepInExPluginsDirectory)" ContinueOnError="true" />
   </Target>
 </Project>
